C_airflow_dags/requirements.txt
apache-airflow==2.7.0
apache-airflow-providers-amazon==8.0.0
boto3==1.26.0
